requirements.txt
absl-py==0.7.1
aiohttp==3.5.4
astor==0.7.1
async-timeout==3.0.1
attrs==19.1.0
boto==2.49.0
boto3==1.9.133
botocore==1.12.133
bz2file==0.98
certifi==2019.3.9
chardet==3.0.4
discord.py==1.0.1
docutils==0.14
gast==0.2.2
gensim==3.7.2
grpcio==1.20.0
h5py==2.9.0
idna==2.8
idna-ssl==1.1.0
jmespath==0.9.4
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
mecab-python3==0.996.1
mock==2.0.0
multidict==4.5.2
numpy==1.16.2
pbr==5.1.3
protobuf==3.7.1
python-dateutil==2.8.0
PyYAML==5.1
requests==2.21.0
s3transfer==0.2.0
scipy==1.2.1
six==1.12.0
smart-open==1.8.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typing-extensions==3.7.2
urllib3==1.24.2
websockets==6.0
Werkzeug==0.15.2
yarl==1.3.0
